% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{aic}
\alias{aic}
\title{Akaike Infomation Criterion for binomial mixture model}
\usage{
aic(mixture)
}
\arguments{
\item{mixture}{estimated mixture model}
}
\value{
The AIC of a model
}
\description{
Akaike Infomation Criterion for binomial mixture model
}

