% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_coverage.R
\name{alleleCounts}
\alias{alleleCounts}
\title{Return allele counts from a genofile based on read counts}
\usage{
alleleCounts(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\value{
an alleleCounts matrix object, which is a list containg the
}
\description{
Return allele counts from a genofile based on read counts
}
\details{
If \code{ref.allele = NULL}, the function returns a list
of matrices corresponding to the read counts of the reference and alternate
alleles respectively, while if \code{ref.allele = 0L} or \code{ref.allele = 1L}
the function returns a  matrix corresponding the reference or alternate
read counts. Currently, the AD tag formats from varscan2re and GATK
are supported.
}
