% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcftools_functions.R
\name{alleleFreqEst}
\alias{alleleFreqEst}
\title{Allele frequency estimate}
\usage{
alleleFreqEst(gdsfile, reference = TRUE)
}
\arguments{
\item{gdsfile}{a SeqVarGDSClass object}

\item{reference}{logical indicating whether to compute the A or B allele
frequency (default TRUE)}
}
\value{
a numeric vector containing allele frequency estimates
}
\description{
Estimate the allele frequency for each variant site
}
\details{
Uses \code{seqApply} to estimate variant frequency over the
SeqVarGDS Class object
}

