% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{autoSelect}
\alias{autoSelect}
\title{Use an information criterion to perform model selection
for number of mixture components}
\usage{
autoSelect(x, N, k, method, ...)
}
\arguments{
\item{x}{vector of observed read counts}

\item{N}{coverage}

\item{k}{maximum number of components to estimate}

\item{method}{string with information criterion either 'bic' or 'aic'}

\item{...}{other parameters passed to binommixEM}
}
\description{
Use an information criterion to perform model selection
for number of mixture components
}

