% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_baf_methods.R
\name{bafMatrix}
\alias{bafMatrix}
\title{Compute B-allele frequency spectrum genome-wide}
\usage{
bafMatrix(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\value{
a bafMatrix object
}
\description{
Compute B-allele frequency spectrum genome-wide
}
\details{
Constructs a bafMatrix object which is a list consisting of 
a matrix of estimated B-allele frequencies and
}

