% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil_functions.R
\name{barcodes}
\alias{barcodes}
\title{Generate GenomicRanges object from GDS file}
\usage{
barcodes(publication)
}
\arguments{
\item{publication}{a character string indicating malaria publication (default Volkmann2008)}
}
\description{
Generate GenomicRanges object from GDS file
}
\details{
Return a GenomicRanges object containing Pf barcode information from Volkmann, 2008.
}

