% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{bic}
\alias{bic}
\title{Bayesian Information Criterion for binomial mixture model}
\usage{
bic(mixture)
}
\arguments{
\item{mixture}{estimated mixture model}
}
\value{
The BIC of a model
}
\description{
Bayesian Information Criterion for binomial mixture model
}

