% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{binommix}
\alias{binommix}
\title{Fit binomial mixture model on coverage data}
\usage{
binommix(y, k, niter = 1000)
}
\arguments{
\item{y}{matrix with 2 columns, first column contains
counts in support of SNV, second in support of ref}

\item{k}{vector of mixture components to fit}

\item{niter}{number of iterations to run}

\item{N}{vector of depth at SNV sites}
}
\description{
Fit binomial mixture model on coverage data
}

