% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{binommixCDF}
\alias{binommixCDF}
\title{CDF for binomial mixture model}
\usage{
binommixCDF(model, x, N)
}
\arguments{
\item{model}{flexmix object}

\item{x}{vector of read counts supporting each SNV}

\item{N}{vector of coverage at SNV}
}
\value{
Vector of theoretical quantiles
}
\description{
CDF for binomial mixture model
}

