% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{binommixEM}
\alias{binommixEM}
\title{EM algorithm for k-component binomial mixture model}
\usage{
binommixEM(x, N, k, mixture.comp = NULL, mixture.weights = NULL,
  epsilon = 1e-06, niter = 1000, verbose = FALSE)
}
\arguments{
\item{x}{observed vector of counts (read counts supporting SNV)}

\item{N}{observed vector of counts (coverage at SNV)}

\item{k}{number of mixture components (max=5)}

\item{mixture.comp}{vector of mixture paramters (length k)}

\item{mixture.weights}{vector of mixture weights (length k)}

\item{epsilon}{precsion (default 1e-6)}

\item{niter}{number of iterations (default 1000)}

\item{verbose}{print running of EM algorithm (default FALSE)}
}
\value{
a list with the following objects
 n total number of SNVs
 k number of mixture components
 cluster.memberships assignments to each class
 cluster.probs an n by k matrix with probability of SNV membership to each class
 pi a k length vector with estimated mixture weights
 mu a k length vector with estimated mixture components
}
\description{
EM algorithm for k-component binomial mixture model
}

