% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{binomMixSlide}
\alias{binomMixSlide}
\title{Fit binomial mixture  model in non-overlapping genomic windows}
\usage{
binomMixSlide(counts_matrix, sample.id, window_list)
}
\arguments{
\item{counts_matrix}{an \code{\link{alleleCounts}} object with ref and alt slots filled}

\item{sample.id}{character sample.id to fit model on.}

\item{window_list}{list of genomic windows}
}
\value{
modelWindow object
}
\description{
Fit binomial mixture  model in non-overlapping genomic windows
}

