% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vcftools_functions.R
\name{bVarFreqEst}
\alias{bVarFreqEst}
\title{Within sample B-allele frequency estimate}
\usage{
bVarFreqEst(gdsfile, sample.id, minDP, minGQ, metadata.granges)
}
\arguments{
\item{gdsfile}{a SeqVarGDS class object}

\item{sample.id}{character string for sample identifier}

\item{minDP}{numeric minimum depth to consider}

\item{minGQ}{numeric minimum geontype quality to consider}

\item{metadata.granges}{a GRanges class object pertaining to variant sites}
}
\value{
a GRanges class object with metadata slot 'bfreq'
}
\description{
Estimate within sample B-allele frequency from variant support
reads and read depth.
}

