% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{callMajor}
\alias{callMajor}
\title{Get major allele calls for isolates from BAF matrix}
\usage{
callMajor(gdsfile, get.nucleotides = FALSE, use.hets = FALSE)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{get.nucleotides}{FALSE logical indicating whether to replace allele
with nucleotide bases}

\item{use.hets}{FALSE logical indicating whether to code hets sepeartely}
}
\value{
a binary matrix  or character matrix containing reference and alternate alleles
}
\description{
Extract major alleles from a GDS file
using the non-reference allele frequency (NRAF) estimated from 
read counts directly. In the case of RAF = 0.5 we select the
major allele randomly from Bernoulli distribution if use.hets is FALSE. Other
wise hets are coded as 3. If get.nucleotides is TRUE a character matrix is
retruned with the alleles coded, otherwise if use.hets the allele is coded as
N. Missing genotype calls are coded as X.
}
