% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcftools_functions.R
\name{chrom_labeller}
\alias{chrom_labeller}
\title{P.falciparum chromosome labeller}
\usage{
chrom_labeller(variable, value)
}
\arguments{
\item{variable}{Pf3D7 chromosome name}

\item{value}{numeric chromsome name}
}
\description{
Key-value pairs for mapping Pf3D7 chromsome names to numeric
chromsome. Used as labeller for facet grid
}

