% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{ciMM}
\alias{ciMM}
\title{Generate Confidence Intervals for
parameter estimates using estimated Fisher Information
Matrix}
\usage{
ciMM(y, model, alpha = 0.05)
}
\arguments{
\item{y}{filtered counts matrix}

\item{model}{flexmix object}

\item{alpha}{confidence level (default = 0.05)}
}
\value{
a data.frame containing parameter estimates
and lower and upper confidence limits
}
\description{
Generate Confidence Intervals for
parameter estimates using estimated Fisher Information
Matrix
}

