% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcftools_functions.R
\name{constructGenotypes}
\alias{constructGenotypes}
\title{Obtain genotypes from a file}
\usage{
constructGenotypes(gdsfile)
}
\arguments{
\item{gdsfile}{a SeqVarGDS class object}
}
\value{
a matrix of genotypes
}
\description{
Contruct matrix of genotypes. Dimension will be nsample * nvar
}

