% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_coverage.R
\name{coverageBySample}
\alias{coverageBySample}
\title{SNP coverage rates by sample thresholds}
\usage{
coverageBySample(coverage_list, threshold)
}
\arguments{
\item{coverage_list}{an \code{\link{alleleCounts}} object}

\item{threshold}{coverage threshold}
}
\value{
a vector of n.snps length
}
\description{
Compute the proportion of samples within each SNP that are covered
to at least threshold.
}

