% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_coverage.R
\name{coverageBySNP}
\alias{coverageBySNP}
\title{Sample coverage rates by SNP thresholds}
\usage{
coverageBySNP(coverage_list, threshold)
}
\arguments{
\item{coverage_list}{an \code{\link{alleleCounts}} object}

\item{threshold}{coverage threshold}
}
\value{
a vector of n.samples length
}
\description{
Compute the proportion of SNPs within each sample that are covered
to at least threshold.
}

