% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truncated_distributions.R
\name{dtbeta}
\alias{dtbeta}
\alias{ptbeta}
\alias{qtbeta}
\alias{rtbeta}
\title{Truncated beta distribution}
\usage{
dtbeta(x, shape1, shape2, lower = 0, upper = 1)

ptbeta(x, shape1, shape2, lower = 0, upper = 1)

qtbeta(p, shape1, shape2, lower = 0, upper = 1)

rtbeta(n, shape1, shape2, lower = 0, upper = 1)
}
\arguments{
\item{x,q}{vector of quantiles}

\item{shape1,shape2}{non-negative parameters of Beta distributions}

\item{lower,upper}{bounds for truncation}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\description{
Density, distribution function, quantile function and
random generation for the truncated Beta distribution with parameters
shape1 and shape2 (and optional truncation bounds)


}
\section{Functions}{
\itemize{
\item \code{ptbeta}: Cumulative distribution function
for truncated beta

\item \code{qtbeta}: Quantile function

\item \code{rtbeta}: Random number generation
}}

