% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_distributions.R
\name{dtexp}
\alias{dtexp}
\alias{lltexp}
\alias{mleTexp}
\title{Truncated exponential distrubition}
\usage{
dtexp(x, theta, lower, upper)

lltexp(x, theta, lower, upper)

mleTexp(x, lower, upper)
}
\arguments{
\item{x}{numeric vector}

\item{theta}{rate parameter for exponential}

\item{lower, upper}{boundary points for truncation}
}
\description{
Density function for random generation for
truncated exponential distribution
}
\section{Functions}{
\itemize{
\item \code{lltexp}: Truncated exponential distrubition

\item \code{mleTexp}: Truncated exponential distrubition
}}

