% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil_functions.R
\name{extractBarcode}
\alias{extractBarcode}
\title{Extract barcode for use in COIL program}
\usage{
extractBarcode(gdsfile, variant.id, barcode.file)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{variant.id}{a numeric vector of variant identifiers}

\item{barcode.file}{character name of output file}
}
\value{
data.frame containing barcode for each sample
}
\description{
Extract barcode for use in COIL program
}
\details{
Generate SNP barcode using supplied variant ids. This
is constructed by collapsing alleles over the variant ids, forming
a barcode for each sample. Then a barcode file containing
sample ids + barcode is generated and can be supplied to the COIL web-app
or to the command line COIL program.
}
\note{
Need to look at how COIL constructed it's orginal barcode. Esp. het calls.
}
\references{
Galinsky, Kevin, et al. "COIL: a methodology for evaluating 
malarial complexity of infection using likelihood from single nucleotide 
polymorphism data." Malaria journal 14.1 (2015): 4.
}
