% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{filterCounts}
\alias{filterCounts}
\title{Filter read-count matrix}
\usage{
filterCounts(x, N)
}
\arguments{
\item{x}{vector of b-allele depths}

\item{N}{vector of site depths}
}
\value{
matrix with two columns
}
\description{
Filter read-count matrix
}

