% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{data}
\name{gambian_isolates}
\alias{gambian_isolates}
\title{Plasmodium falciparum SNV data from Gambia.}
\format{A \code{\link[SeqArray]{SeqVarGDSClass}} object.}
\source{
\url{ftp://ngs.sanger.ac.uk/production/pf3k/release_4/}
}
\description{
Variant calls of polymorphic variants from 65 P.falciparum isolates in The Gambia,
from Amambua-Ngwa A, Tetteh KK et al., 2012 \url{http://www.ncbi.nlm.nih.gov/pubmed/23133397}.
The data was extracted from VCF files and converted to GDS format from Malaria Genomics Consortium Pf3k version 4. See the README file
at \url{ftp://ngs.sanger.ac.uk/production/pf3k/release_4/} for details. The GDS object was obtained
by filtering the MalariaGen VCF files and keeping only biallelic SNPs that were polymorphic in The Gambia
and passed all filters.
}

