% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_baf_methods.R
\name{generateWindows}
\alias{generateWindows}
\title{Construct non-overlapping Genomic Windows by Chromosome}
\usage{
generateWindows(gdsfile, window_size)
}
\arguments{
\item{gdsfile}{a \code{\link{SeqVarGDSClass}} object}

\item{window_size}{size of overlapping window in base-pairs}
}
\description{
Construct non-overlapping Genomic Windows by Chromosome
}
