% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_baf_methods.R
\name{getBAFvar}
\alias{getBAFvar}
\title{Estimate variance in BAF spectra along the genome in non-overlapping windows}
\usage{
getBAFvar(baf_matrix, window_size, by.sample = FALSE)
}
\arguments{
\item{baf_matrix}{a \code{\link[moimix]{bafMatrix}} object}

\item{window_size}{integer size of window in bp}

\item{by.sample}{FALSE partition by sample}
}
\value{
data.frame with chromosome, window id, start, midpoint and end of window
and estimates of average variance for window. If by.sample is TRUE, then there
will be additional sample.id column with
}
\description{
Estimate variance in BAF spectra along the genome in non-overlapping windows
}
\details{
This function computes
}

