% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil_functions.R
\name{getBarcodeVariants}
\alias{getBarcodeVariants}
\title{Helper functions for extracting variant IDs corresponding to previously published
barcodes.}
\usage{
getBarcodeVariants(gdsfile, publication = "Volkman2008")
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{publication}{a character string representing a publication}
}
\description{
Helper functions for extracting variant IDs corresponding to previously published
barcodes.
}
