% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Helper function for producing data frame of genomic coordinates}
\usage{
getCoordinates(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\description{
Helper function for producing data frame of genomic coordinates
}
