% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{getDosage}
\alias{getDosage}
\title{Helper function for producing matrix of allele dosages}
\usage{
getDosage(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\description{
Helper function for producing matrix of allele dosages
}
