% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{getGATKInfo}
\alias{getGATKInfo}
\title{Get GATK info tags}
\usage{
getGATKInfo(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\value{
a data.frame with the following columns:
 chromosome
 position
 variant.id
 qd - (average quality by depth)
 mq - RMS mapping quality
 fs - Fisher strand bias
 mqrank - MQ rank
 haplotypescore
 readposranksum
}
\description{
Obtain GATK info tags relevant to best practices.
}
\references{
\url{http://gatkforums.broadinstitute.org/gatk/discussion/2806/howto-apply-hard-filters-to-a-call-set}
}
