% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_popgen.R
\name{getHeterozygosity}
\alias{getHeterozygosity}
\title{Compute population level heterozygosity}
\usage{
getHeterozygosity(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\description{
Compute population level heterozygosity
}
\note{
Need to update so population ids can be submitted.
}

