% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_popgen.R
\name{getHeterozygosityBySample}
\alias{getHeterozygosityBySample}
\title{Compute heterozygosity by sample}
\usage{
getHeterozygosityBySample(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\value{
a numeric matrix of size l by n where l is the number of samples
and n is the number of SNPs.
}
\description{
Compute heterozygosity by sample
}
\details{
Follows the method outlined in Manske et al. 2012. Briefly,
for each sample, the reference and and alternate allele frequencies are computed
as the proportion of reads covering each allele. Then heterozygosity at a SNP
is 1 - (raf^2 + aaf^2)
}
