% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_popgen.R
\name{getMAF}
\alias{getMAF}
\title{Compute minor allele frequency}
\usage{
getMAF(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\description{
Compute minor allele frequency
}
\details{
Coverage based estimation of minor allele frequences.
The MAF  is computed by computing the minimum of the proportion of 
reads covering the reference and alternate allele over all samples.
}
\note{
Currently only supports gds files obtained from GATK callers
}
