% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{getMSE}
\alias{getMSE}
\title{Model assessment for fitted model using MSE}
\usage{
getMSE(fitted_models)
}
\arguments{
\item{fitted_models}{a moimix object}
}
\description{
Model assessment for fitted model using MSE
}

