% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{getTheta}
\alias{getTheta}
\title{Return estimated model parameters}
\usage{
getTheta(model, k = NULL, criterion = NULL)
}
\arguments{
\item{model}{stepFlexmix or flexmix object}

\item{k}{if stepFlexmix choose model with k components}

\item{criterion}{if stepFlexmix choose model according to information 
criterion}
}
\description{
Return estimated model parameters
}

