% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{data}
\name{global.allele.freq}
\alias{global.allele.freq}
\title{Minor Allele Frequenies of 86,000 round Plasmodium falciparum SNPs.}
\format{A data frame with 86158 rows and 11 variables:
\describe{
  \item{ID}{Row number}
  \item{Chromosome}{Chromosome}
  \item{Position}{Physical position of SNP}
  \item{snpName}{SNP identifier}
  \item{Geme}{Gene name}
  \item{MAF_AFR}{Minor allele frequency in Africa}
  \item{MAF_SEA}{Minor allele frequency in South East Asia}
  \item{MAF_PNG}{Minor allele frequency in Papua New Guinea}
  \item{GeneID}{Gene Identifier}
  \item{GeneAliases}{Alternate Gene Identifier}
  \item{GeneDescription}{Proposed function of gene}


}}
\source{
\url{http://www.malariagen.net/data}
}
\description{
A dataset containing information about exonic SNPs Plasmodium falciparum
from Manske et al., 2012 \url{http://www.ncbi.nlm.nih.gov/pubmed/22722859}.
}

