% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/haldane.R
\docType{package}
\name{haldane}
\alias{haldane}
\alias{haldane-package}
\title{Analayse whole-genome sequencing data from malaria parasites}
\description{
\pkg{haldane} enables population genetics analysis on Variant Call Format
(VCF) files obtained from WGS experiments.
}

