% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moi_simulation.R
\name{handleMOI}
\alias{handleMOI}
\title{Apply handling strategies to simulated data}
\usage{
handleMOI(simMOI.list)
}
\arguments{
\item{simMOI.list}{The list output from simulateMOI}
}
\value{
A list containing three entries:
 true.alleles: A matrix containing the true alleles from simulation
 monoclonal.alleles: A matrix with MOI=1 samples
 sample.major.allels: A matrix containing major alleles for each
 sample and SNV
}
\description{
Apply handling strategies to simulated data
}
\details{
This function generates several datasets pertaining
to different mulitplicity of infection strategies. Strategy A
simply removes all samples with MOI > 1. Strategy B calls the
major allele for each sample and single nucleotide variant.
}

