% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{harmonicMean}
\alias{harmonicMean}
\title{Harmonic Mean}
\usage{
harmonicMean(array)
}
\arguments{
\item{array}{numeric vector}
}
\description{
This function returns the harmonic mean of an array.
Code from http://economistatlarge.com/r-guide/arithmetic-harmonic-geometric-means-r
}
\examples{
harmonicMean(rep(100,10))
}
\keyword{harmonic}
\keyword{mean}

