% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{initEM}
\alias{initEM}
\title{Initialise mixture model parameters for EM
uses random start by default}
\usage{
initEM(x, N, k, method = "random_seed")
}
\description{
Initialise mixture model parameters for EM
uses random start by default
}

