% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{kmeans_seed}
\alias{kmeans_seed}
\title{Seeding methods for mixture model
kmeans initilisation}
\usage{
kmeans_seed(x, N, k)
}
\description{
Seeding methods for mixture model
kmeans initilisation
}

