% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{data}
\name{maf_summary}
\alias{maf_summary}
\title{Minor Allele Frequenies of ~ 30,000 Plasmodium falciparum SNPs.}
\format{
A data frame with 29621 rows and 10 variables:
}
\source{
\url{http://www.malariagen.net/data}
}
\description{
A dataset containing information about exonic SNPs in Plasmodium falciparum
from Manske et al., 2012 \url{http://www.ncbi.nlm.nih.gov/pubmed/22722859}.
}
