% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{mixLL}
\alias{mixLL}
\title{log-Likelihood function for mixture model
Expected log-Likelihood function for mixture model}
\usage{
mixLL(x, N, k, class.probs, mixture.comp, mixture.weights)
}
\description{
log-Likelihood function for mixture model
Expected log-Likelihood function for mixture model
}

