% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moimix.R
\docType{package}
\name{moimix}
\alias{moimix}
\title{Evaluate muliplicity of infection in malaria parasites}
\description{
\pkg{moimix} enables the estimation of MOI from SNV data and to evaluate
the effect of MOI on SNV data.
}
