% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{mse}
\alias{mse}
\title{Mean Square Error for binomial mixture model}
\usage{
mse(model, theta)
}
\arguments{
\item{theta}{true mixture-model parameters of form (pi_1,..,pi_k, mu_1,...mu_k)}

\item{mixture}{estimated mixture model}
}
\description{
Compute sum of squares for mixture components
}

