% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{mseMM}
\alias{mseMM}
\title{Mean Square Error for binomial mixture model}
\usage{
mseMM(mixture, pi)
}
\arguments{
\item{mixture}{estimated mixture model}

\item{pi}{true mixture weights}
}
\description{
Mean Square Error for binomial mixture model
}

