% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{perSiteAlleleCoverage}
\alias{perSiteAlleleCoverage}
\title{Compute per site per sample coverage}
\usage{
perSiteAlleleCoverage(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\value{
a matrix of allele counts
}
\description{
Compute per site per sample coverage
}
