% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{perSiteCoverage}
\alias{perSiteCoverage}
\title{Compute per variant depth accross all samples.}
\usage{
perSiteCoverage(gdsfile)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}
}
\value{
an integer vector
}
\description{
Compute per variant depth accross all samples.
}

