% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{perSiteCoverageBySample}
\alias{perSiteCoverageBySample}
\title{Compute the proportion of samples covered up any number of bases}
\usage{
perSiteCoverageBySample(gdsfile, threshold)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{threshold}{integer lower bound for coverage}
}
\value{
a vector of length equal to number of variants with
the proportion of samples covered up to threshold
}
\description{
Compute the proportion of samples covered up any number of bases
}

