% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{perSiteCoverageBySNP}
\alias{perSiteCoverageBySNP}
\title{Compute proportion of SNPs covered up to a lower bound of bases within 
each sample.}
\usage{
perSiteCoverageBySNP(gdsfile, threshold)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{threshold}{integer lower bound for coverage}
}
\value{
a vector of length equal to number of samples with
the proportion of SNPs covered up to threshold
}
\description{
Compute proportion of SNPs covered up to a lower bound of bases within 
each sample.
}

