% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vcftools_functions.R
\name{plotVarFreq}
\alias{plotVarFreq}
\title{Plot B-allele frequency genome wide}
\usage{
plotVarFreq(sample.granges)
}
\arguments{
\item{sample.granges}{a GRanges object}
}
\value{
a ggplot2 object
}
\description{
Plot allele-frequency spectrum accross genome and facet by
chromosome
}

