% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{data}
\name{png.allele.frq}
\alias{png.allele.frq}
\title{Minor allele frequencies of 110,000 PNG Plasmodium falciparum SNPs}
\format{A data.frame with 107879 rows and 5 columns
\describe{
 \item{chr}{Pf3D7 chromsome name}
 \item{position}{Physical position}
 \item{a.frq}{Reference allele frequency}
 \item{b.frq}{Alternate allele frequency}
 \item{maf}{Minor allele frequency}
}}
\description{
A dataset containing SNP frequencies derived from 104 isolates
from Papua New Guinea
}

