% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_distributions.R
\name{qtbeta}
\alias{qtbeta}
\title{@describeIn dtbeta Quantile function
 @param p vector of probabilities}
\usage{
qtbeta(p, shape1, shape2, lower = 0, upper = 1)
}
\description{
@describeIn dtbeta Quantile function
 @param p vector of probabilities
}

