% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{random_seed}
\alias{random_seed}
\title{random start seeding
choose array of starting points choose iteration that
gives best estimate}
\usage{
random_seed(x, N, k, nstart = 20)
}
\description{
random start seeding
choose array of starting points choose iteration that
gives best estimate
}

