% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rdirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution random number generator}
\usage{
rdirichlet(a)
}
\arguments{
\item{a}{vector of concentration parameters}
}
\value{
length(a) dirichlet random numbers
}
\description{
Draw random numbers from a Dirichlet distribution with parameters given by the vector a.
Code modified from http://stats.stackexchange.com/a/30972
}
\examples{
rdirichlet(c(1,2,4))
}
\keyword{Dirichlet}
\keyword{distribution}

