% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coil_functions.R
\name{readEstMOI}
\alias{readEstMOI}
\title{Parse estMOI log files in a directory into tidy data frame}
\usage{
readEstMOI(path)
}
\arguments{
\item{path}{path containing .txt files containing estMOI estimates}
}
\value{
a data.frame
}
\description{
Parse estMOI log files in a directory into tidy data frame
}
\note{
Assumes that outPrefix set as option in estMOI is set to sample identifer.
}
