% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vcftools_functions.R
\name{sampleMissingness}
\alias{sampleMissingness}
\title{Missing rate per sample}
\usage{
sampleMissingness(gdsfile)
}
\arguments{
\item{gdsfile}{a SeqVarGDSClass object}
}
\value{
a numeric vector containing proporition of missing calls in each
sample
}
\description{
Calculate the proportion of "No calls" in a sample
}
\details{
Uses \code{seqApply} to compute the per sample missingness rate
from a SeqVarGDSClass object
}

