% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{seMM}
\alias{seMM}
\title{Compute standard errors for mixture model estimates}
\usage{
seMM(x, N, mixture)
}
\arguments{
\item{x}{vector of read counts}

\item{N}{vector of coverage at site}
}
\description{
se is computed as the inverse of the Fisher information matrix
at the EM estimates
}

