% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_moi_simulation.R
\name{simulate_moi}
\alias{simulate_moi}
\title{Simple simulation of MOI SNV data}
\usage{
simulate_moi(n.samples, n.snps, moi, coverage, pi.true = NULL,
  mu.true = NULL, aaf.dist, ...)
}
\arguments{
\item{n.samples}{number of infected hosts in population}

\item{n.snps}{number of SNPs observed}

\item{moi}{number of clonal infections}

\item{coverage}{vector of total coverage per sample}

\item{aaf.dist}{sampling distribution for SNV frequencies}

\item{...}{other parameters to pass to aaf.dist}

\item{clonal.props}{optional matrix of true mixture proportions}

\item{geno.props}{optional matrix of true mixture components}
}
\value{
A list containing the following elements
 pi.true an n.samples by moi matrix containing true clonal proportions
 mu.true an n.samplps by moi matrix containing true genotype proportions
 aaf an n.snps vector of under SNV probabilities
 reads an n.samples by n.snps matrix containing read counts supporting each SNV
}
\description{
Generate read count data in support of SNVs partioned by
mixtures of clones within a host. This is done by setting true
clonal mixture proportions and the underlying sampling probabilities
for each clone. Then a random sample of underlying SNV probabilities
is sampled from either a rbeta or rtexp distrubition with given parameters
and read counts are generated by bionimial conditonal on the underlying
SNV probability with the clone sampling probabilities.
}

