% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moi_simulation.R
\name{simulateMOI}
\alias{simulateMOI}
\title{Simulate multiplicity of infection in parasite sequencing data}
\usage{
simulateMOI(n.samples, n.snps, moi, coverage, error, clone.props = NULL,
  maf.dist = rtexp, ...)
}
\arguments{
\item{n.samples}{integer number of humans sampled for Plasmodium infection}

\item{n.snps}{integer number of SNPs examined}

\item{moi}{integer number of infections per human}

\item{coverage}{integer number of reads covering each base for each isolate}

\item{error}{numeric probability of error in base call}

\item{clone.props}{optional n.samples by moi matrix of fixed clonal proportions (default NUL)}

\item{maf.dist}{minor allele frequency distribution sampling function
(DEFAULT=rtexp)}

\item{...}{parameters to be passed to maf.dist See details.}
}
\value{
A list with the following elements
 clone.props: a matrix of dimension n.samples by moi
 allele.freq: A vector of length n.snps containing observed allele frequency
 distribution generated by maf.dist
 obs.error.prop: the total observed error proportion
 obs.alt.counts: A matrix of size n.samples by n.snps containing the
 observed number of reads in support of each SNP.
}
\description{
Simulate multiplicity of infection in parasite sequencing data
}
\details{
This function simulates SNV data for fixed MOI.
It requires passing a function as an argument that specifies the
alternate allele frequency distribution.
The sampling distributions available are either the truncated exponential
or beta. The parameters for these distributions are passed as additional
arguments to the function. For rtexp a mutation rate (theta) and truncation
parameter t are required, while for rbeta shape parameters are required.
}

