% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{updateClassProb}
\alias{updateClassProb}
\title{Update class probabilities (i.e compute tau in Estep)}
\usage{
updateClassProb(x, N, k, mixture.comp, mixture.weights)
}
\description{
Update class probabilities (i.e compute tau in Estep)
}

