% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_coverage.R
\name{alleleCounts}
\alias{alleleCounts}
\title{Return allele counts from a genofile based on read counts}
\usage{
alleleCounts(gdsfile, ref.allele = NULL)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{ref.allele}{\code{NULL} a single numeric value.}
}
\description{
Return allele counts from a genofile based on read counts
}
\details{
If \code{ref.allele = NULL}, the function returns a list
of matrices corresponding to the read counts of the reference and alternate
alleles respectively, while if \code{ref.allele = 0L} or \code{ref.allele = 1L}
the function returns a  matrix corresponding the reference or alternate
read counts. Currently, the variant formats from varscan and GATK
are supported.
}

