% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{binommix}
\alias{binommix}
\title{Fit binomial mixture model on coverage data}
\usage{
binommix(counts_matrix, sample.id, k, coverage_threshold = 0L, niter = 1000)
}
\arguments{
\item{counts_matrix}{an \code{\link{alleleCounts}} object with ref and alt slots filled}

\item{sample.id}{character sample.id to fit model on.}

\item{k}{vector of mixture components to fit}

\item{coverage_threshold}{exclude sites with total coverage < coverage_threshold}

\item{niter}{number of iterations to run}
}
\description{
Fit binomial mixture model on coverage data
}
