% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{collapseProbs}
\alias{collapseProbs}
\title{Collapse allele frequency to half axis}
\usage{
collapseProbs(p)
}
\arguments{
\item{p}{vector of proprotions}
}
\description{
Collapse allele frequency to half axis
}

