% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truncated_distributions.R
\name{dtexp}
\alias{dtexp}
\title{Truncated exponential distrubition}
\usage{
dtexp(x, theta, lower, upper)
}
\arguments{
\item{x}{numeric vector}

\item{theta}{rate parameter for exponential}

\item{lower,}{upper boundary points for truncation}
}
\description{
Density function for random generation for
truncated exponential distribution
}

