% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_helpers.R
\name{extractPED}
\alias{extractPED}
\title{Extract PED files from gds}
\usage{
extractPED(gdsfile, use.hets = FALSE, out.file)
}
\arguments{
\item{gdsfile}{a \code{\link[SeqArray]{SeqVarGDSClass}} object}

\item{use.hets}{FALSE include heterozygote genotypes}

\item{out.file}{prefix of PLINK files for output}
}
\description{
Extract PED files from gds
}
\details{
This function writes a plink .ped and .map file for a given
gdsfile. If the use.hets option is true the genotypes are used as is, other
wise heterzygotes are set to missing.
}

