% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{filterCounts}
\alias{filterCounts}
\title{Filter read-count matrix}
\usage{
filterCounts(x, N, p = c(0.01, 0.99))
}
\arguments{
\item{x}{vector of b-allele depths}

\item{N}{vector of site depths}

\item{p}{vector containing boundary cut-offs for inclusion}
}
\value{
matrix with two columns
}
\description{
Filter read-count matrix
}

