% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moimix.R
\docType{package}
\name{haldane}
\alias{haldane}
\alias{haldane-package}
\title{Evaluate muliplicity of infection in malaria parasites}
\description{
\pkg{moimix} enables the estimation of MOI from SNV data and to evaluate
the effect of MOI on SNV data.
}

