% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{infoMat}
\alias{infoMat}
\title{Fisher Information Approximation}
\usage{
infoMat(x, N, k, mu, pi)
}
\arguments{
\item{x}{vector of read counts}

\item{N}{vector of coverage at site}

\item{k}{number of components}

\item{mu}{mixture components}

\item{pi}{mixture weights}
}
\value{
A 2k by 2k Fisher information matrix
}
\description{
Estimate the Fisher information matrix for
a bionimal mixture model
}

