% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{infomat}
\alias{infomat}
\title{Fisher Information Approximation}
\usage{
infomat(y, model)
}
\arguments{
\item{y}{a two column matrix of read counts}

\item{model}{flexmix object}
}
\value{
A 2k by 2k Fisher information matrix
}
\description{
Estimate the Fisher information matrix for
a bionimal mixture model
}
