% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture.R
\name{initEM}
\alias{initEM}
\title{#return pdf of two component binomial mixture
Initialise mixture model parameters for EM
uses kmeans on proportions}
\usage{
initEM(x, N, k)
}
\description{
#return pdf of two component binomial mixture
Initialise mixture model parameters for EM
uses kmeans on proportions
}

