% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{mseMM}
\alias{mseMM}
\title{Mean Square Error for binomial mixture model}
\usage{
mseMM(mixture, theta)
}
\arguments{
\item{mixture}{estimated mixture model}

\item{theta}{true mixture-model parameters of form (pi_1,..,pi_k, mu_1,...mu_k)}
}
\description{
Mean Square Error for binomial mixture model
}

