% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_baf_methods.R
\name{plot.bafMatrix}
\alias{plot.bafMatrix}
\title{Plot bafMatrix object}
\usage{
plot.bafMatrix(baf, sample.id, assignments = NULL, ...)
}
\arguments{
\item{baf}{a bafMatrix object}

\item{sample.id}{character name of sample to plot}

\item{assignments}{integer vector of cluster memberships (NULL)}
}
\description{
Plot bafMatrix object
}
\details{
plots the genome-wide signal of MOI within an isolate
}

