% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_distributions.R
\name{rtexp}
\alias{rtexp}
\title{Truncated exponential distribution random number generation}
\usage{
rtexp(n, theta, t)
}
\arguments{
\item{n}{number of random values to return}

\item{t}{cut-off threshold for truncation}

\item{mu}{rate parameter for exponential}
}
\value{
n random values from TEXP
}
\description{
Simulate random numbers from a truncated exponential distribution
}
\details{
Simulate SNP alternate allele frequencies from a truncated exponential distribution
Code is modified from that given at https://stat.ethz.ch/pipermail/r-help/2010-June/242615.html
}

