% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utility.R
\name{sampleMM}
\alias{sampleMM}
\title{Generate iid mixture random variables 
for testing EM implementation}
\usage{
sampleMM(n, N, k, mu, pi)
}
\arguments{
\item{n}{number of realisations}

\item{N}{number of trials}

\item{k}{number of components}

\item{mu}{true mixture components}

\item{pi}{true mixture weights}
}
\description{
Generate iid mixture random variables 
for testing EM implementation
}

