% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqarray_process_coverage.R
\name{scaledProportion}
\alias{scaledProportion}
\title{sum number of samples/snps meeting coverage threshold scaled by total number
of SNPs/proportions}
\usage{
scaledProportion(x, threshold, scale)
}
\arguments{
\item{x}{integer vector of read counts}

\item{threshold}{integer lower bound for coverage}

\item{scale}{integer to scale results.}
}
\description{
sum number of samples/snps meeting coverage threshold scaled by total number
of SNPs/proportions
}

