% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_mixture_utility.R
\name{seMM}
\alias{seMM}
\title{Compute standard errors for mixture model parameter estimates}
\usage{
seMM(y, model)
}
\arguments{
\item{y}{a two column matrix of read counts}

\item{model}{flexmix object}
}
\description{
se is computed as the inverse of the Fisher information matrix
at the EM estimates
}

