% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/moi_simulation.R
\name{simulateMOI}
\alias{simulateMOI}
\title{Simulate multiplicity of infection in parasite sequencing data}
\usage{
simulateMOI(n.samples, n.snps, moi, clone.props = NULL, coverage, error,
  dirichlet.param, maf.dist = rtexp, ...)
}
\arguments{
\item{n.samples}{integer number of humans sampled for Plasmodium}

\item{n.snps}{integer number of SNPs examined}

\item{moi}{integer number of infections per human}

\item{coverage}{integer number of reads covering each base for each isolate}

\item{error}{numeric probability of error in base call}

\item{dirichlet.param}{integer vector of length 2, specifying mixing
parameters for dirichilet distribution}

\item{maf.dist}{minor allele frequency distribution sampling function
(DEFAULT=rtexp)}

\item{...}{parameters to be passed to maf.dist See details.}
}
\value{
A list with the following elements
 clone.props: a list of n.sample vectors of length K containing the proprotion
 of each clone for each sample
 allele.freq: A vector of length n.snps containing observed allele frequency
 distribution generated by maf.dist
 obs.error.prop: the total observed error proportion
 obs.alt.counts: A matrix of size n.samples by n.snps containing the
 observed number of reads in support of each SNP.
}
\description{
Simulate multiplicity of infection in parasite sequencing data
}
\details{
This function simulates MOI data and applies two simple MOI
handling strategies. It assumes that MOI in a population is fixed.
It requires passing a function as
an argument that specifies the minor allele frequency distribution.
The sampling distributions available are either truncated exponential
or beta. The parameters for these distributions are passed as additional
arguments to the function. For rtexp a mutation rate and truncation
parameter are required, while for rbeta shape parameters are required.
}

