% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_moi_simulation.R
\name{simulateMOI}
\alias{simulateMOI}
\title{Simple simulation of MOI SNV data}
\usage{
simulateMOI(
  n.samples,
  n.snps,
  moi,
  mean_coverage,
  error_coverage,
  pi.true = NULL,
  mu.true = NULL
)
}
\arguments{
\item{n.samples}{number of infected hosts in population}

\item{n.snps}{number of SNPs observed}

\item{moi}{number of clonal infections}

\item{mean_coverage}{average coverage of reads}

\item{error_coverage}{deviation in coverage of reads}

\item{pi.true}{optional matrix of true mixture proportions}

\item{mu.true}{optional matrix of true mixture components}
}
\value{
A list containing the following elements
 pi.true an n.samples by moi matrix containing true clonal proportions
 mu.true an n.samplps by moi matrix containing true genotype proportions
 aaf an n.snps vector of under SNV probabilities
 read.counts an n.samples by n.snps matrix containing read counts supporting each SNV
 error.counts an n.samples by n.snps matrix containing number of error reads
}
\description{
Generate read count data in support of SNVs partioned by 
mixtures of clones within a host. This is done by setting true
clonal mixture proportions and the underlying sampling probabilities
for each clone. Then a random sample of underlying SNV probabilities
is sampled from either a rbeta or rtexp distrubition with given parameters
and read counts are generated by bionimial conditonal on the underlying
SNV probability with the clone sampling probabilities.
}
