% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateMOI}
\alias{simulateMOI}
\title{Simulate multiplicity of infection in parasite sequencing data}
\usage{
simulateMOI(n.samples, n.snps, moi.lambda, coverage, error, dirichlet.param,
  maf.dist = rtexp, ...)
}
\arguments{
\item{n.samples}{integer number of humans sampled for Plasmodium}

\item{n.snps}{integer number of SNPs examined}

\item{coverage}{integer number of reads covering each base}

\item{error}{numeric probability of error in base call}

\item{dirichlet.param}{integer vector of length 2, specifying mixing
parameters for dirichilet distribution}

\item{maf.dist}{minor allele frequency distribution sampling function
(DEFAULT=rtexp)}

\item{...}{parameters to be passed to maf.dist See details.}

\item{moi.labmda}{integer expected number of infections per human}
}
\value{
A list with the following elements
 param: a list containing arguments passed to simulateMOI
 moi.observed: a vector containing true MOI for each sample
 dirichlet.alphas: a vector containing concentration parameters passed
 to rdirichlet.
 clone.props: a list of n.sample vectors of length K containing the proprotion
 of each clone for each sample
 reads.per.clone: A list of n.sample matrices of n.snp by K size containing
 number of reads observed at each SNP for each clone.
 allele.freq: A vector of length n.snps containing observed allele frequency
 distribution generated by maf.dist
 alt.counts: A matrix of size n.samples by n.snps size containing
 number of reads in support of alternate allele.
 error.counts: A list of n.sample matrices of K by n.snps size containing
 number of sequencing errors per clone.
 obs.error.prop: the total observed error proportion
 error.alt.change: A matrix of size n.samples by n.snps containing
 effect of sequence error on SNP calls
 obs.alt.counts: A matrix of size n.samples by n.snps containing the
 observed number of reads in support of each SNP.
}
\description{
Simulate multiplicity of infection in parasite sequencing data
}
\details{
This function simulates MOI data and applies two simple MOI
handling strategies. It assumes that MOI in a population is distributed
according to a Poisson distribution. It requires passing a function as
an argument that specifies the minor allele frequency distribution.
The sampling distributions available are either truncated exponential
or beta. The parameter for these distributions are passed as additional
arguments to the
}

